glfw for windows, vulkan for 32bit and 64bit, eng_Array type, and more.

+++ Project/Engine/Ini.h
@@ -5,6 +5,7 @@ extern "C" {
 #else
 #include <stdbool.h>
 #endif
+#include <stdint.h>
 
 	////////////////////////////////////////////////////////////////////////// Lifecycle
 	/**
@@ -37,7 +38,7 @@ extern "C" {
 	 * @return the sizeof the internal eng_Url object, for use with custom
 	 * allocators.
 	 */
-	unsigned eng_IniRGetSizeof();
+	size_t eng_IniRGetSizeof();
 
 	////////////////////////////////////////////////////////////////////////// Ini API
